internal references
