+++ apps/website/app/api/supabase/space/route.ts
@@ -1,5 +1,4 @@
 import { NextResponse, NextRequest } from "next/server";
-import { createClient } from "~/utils/supabase/server";
 import {
   createApiResponse,
   handleRouteError,
