+++ packages/database/supabase/functions/create-space/deno.json
@@ -1,6 +1,11 @@
 {
   "imports": {
-    "runtime": "jsr:@supabase/functions-js/edge-runtime.d.ts",
-    "supabase": "jsr:@supabase/supabase-js@2"
+    "@supabase/functions-js/edge-runtime": "jsr:@supabase/functions-js/edge-runtime.d.ts",
+    "@supabase/supabase-js": "jsr:@supabase/supabase-js@2",
+    "@supabase/functions-js": "jsr:@supabase/functions-js@2",
+    "@repo/database/types.gen": "../../../types.gen.ts",
+    "@repo/ui/lib/supabase/contextFunctions": "../../../../ui/src/lib/supabase/contextFunctions.ts",
+    "@repo/ui/lib/supabase/client": "../../../../ui/src/lib/supabase/client.ts",
+    "@repo/ui/lib/execContext": "../../../../ui/src/lib/execContext.ts"
   }
 }
